requirements.txt
streamlit~=1.41.1
torch~=2.5.1
torchvision~=0.20.1
pillow~=10.4.0
numpy~=1.26.4
matplotlib~=3.10.0
torchsummary~=1.5.1